refactor: 비대해진 detail screen의 위젯들 분리
