+++ lib/features/workout_history/presentation/widgets/workout_detail_map_widget.dart
@@ -1,0 +1,96 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_map/flutter_map.dart';
+import 'package:latlong2/latlong.dart';
+import 'package:ridingmate/core/extensions/theme_extensions.dart';
+import 'package:ridingmate/features/workout_history/domain/entities/location_data.dart';
+import 'package:ridingmate/features/workout_history/domain/entities/workout_record.dart';
+import 'package:ridingmate/shared/design_system/tokens/semantic_colors.dart';
+import 'package:ridingmate/shared/map/common_map_widgets.dart';
+import 'package:ridingmate/shared/map/map_constants.dart';
+
+//TODO: 추후 api 연결 시 폴리곤 띄우기 방식 변경, bbox로 지도 크기 조정
+class WorkoutDetailMapWidget extends StatelessWidget {
+  const WorkoutDetailMapWidget({super.key, required this.workoutRecord});
+
+  final WorkoutRecord workoutRecord;
+
+  static const LatLng _defaultCenter = MapConstants.seoulCityHall;
+  static const double _defaultZoom = MapConstants.defaultZoom;
+
+  @override
+  Widget build(BuildContext context) {
+    final SemanticColors colors = context.semanticColor;
+
+    final List<LatLng> routePoints = _convertLocationDataToLatLng(
+      workoutRecord.locationData,
+    );
+
+    // LatLngBounds를 사용한 카메라 설정
+    final CameraFit? cameraFit = _calculateCameraFit(routePoints);
+
+    return FlutterMap(
+      options: MapOptions(
+        initialCenter: _defaultCenter,
+        initialZoom: _defaultZoom,
+        initialCameraFit: cameraFit,
+        interactionOptions: const InteractionOptions(
+          flags: InteractiveFlag.none,
+        ),
+      ),
+      children: <Widget>[
+        CommonMapWidgets.createTileLayer(),
+        // 운동 경로 폴리라인
+        if (routePoints.isNotEmpty)
+          PolylineLayer<LatLng>(
+            polylines: <Polyline<LatLng>>[
+              Polyline<LatLng>(
+                points: routePoints,
+                color: colors.primaryNormal,
+                strokeWidth: MapConstants.polylineStrokeWidth,
+              ),
+            ],
+          ),
+        CommonMapWidgets.createAttributionWidget(),
+      ],
+    );
+  }
+
+  /// WorkoutRecord의 locationData를 LatLng 포인트들로 변환 TODO : 추후 mapper에서 변환
+  List<LatLng> _convertLocationDataToLatLng(List<LocationData> locationData) {
+    return locationData
+        .map((LocationData data) => LatLng(data.latitude, data.longitude))
+        .toList();
+  }
+
+  CameraFit? _calculateCameraFit(List<LatLng> routePoints) {
+    if (routePoints.isEmpty) {
+      return null;
+    }
+
+    final LatLngBounds bounds = _calculateLatLngBounds(routePoints);
+    return CameraFit.bounds(bounds: bounds, padding: const EdgeInsets.all(20));
+  }
+
+  LatLngBounds _calculateLatLngBounds(List<LatLng> points) {
+    if (points.isEmpty) {
+      throw ArgumentError('Points list cannot be empty');
+    }
+
+    double minLat = points.first.latitude;
+    double maxLat = points.first.latitude;
+    double minLng = points.first.longitude;
+    double maxLng = points.first.longitude;
+
+    for (final LatLng point in points) {
+      if (point.latitude < minLat) minLat = point.latitude;
+      if (point.latitude > maxLat) maxLat = point.latitude;
+      if (point.longitude < minLng) minLng = point.longitude;
+      if (point.longitude > maxLng) maxLng = point.longitude;
+    }
+
+    return LatLngBounds(
+      LatLng(minLat, minLng), // southwest
+      LatLng(maxLat, maxLng), // northeast
+    );
+  }
+}
